Changed the drawing part of PianoRoll to be implemented by class
